Explore codes and lists

Used to get details about codes and code lists stored by ONS. Codes are used to provide a common definition when presenting statistics with related categories. Codes are gathered in code lists, which may change over time to include new or different codes. The meaning of a code should not change over time, but new codes may be created where new meaning is required.

ons_codes(id = NULL, edition = NULL)

ons_code(id = NULL, edition = NULL, code = NULL)

ons_code_dataset(id = NULL, edition = NULL, code = NULL)

Arguments

id
[character] Id that represents a dataset.

edition
[character] A subset of the dataset representing a specific time period. For some datasets this edition can contain all time periods (all historical data). The latest version of this is displayed by default.

code
[character] The ID of the code within a code list.

Value

A list or character vector.

Examples

# \donttest{
#codes
ons_codes(id = "quarter", edition = "one-off")

#>   code label                                        links.href
#> 1   q1    Q1 https://api.beta.ons.gov.uk/v1/code-lists/quarter
#> 2   q2    Q2 https://api.beta.ons.gov.uk/v1/code-lists/quarter
#> 3   q3    Q3 https://api.beta.ons.gov.uk/v1/code-lists/quarter
#> 4   q4    Q4 https://api.beta.ons.gov.uk/v1/code-lists/quarter
#>                                                                             links.href
#> 1 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q1/datasets
#> 2 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q2/datasets
#> 3 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q3/datasets
#> 4 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q4/datasets
#>                                                               links.self.href
#> 1 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q1
#> 2 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q2
#> 3 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q3
#> 4 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q4
#>   links.self.id
#> 1            q1
#> 2            q2
#> 3            q3
#> 4            q4

ons_code(id = "quarter", edition = "one-off", code = "q2")

#> $code
#> [1] "q2"
#> 
#> $label
#> [1] ""
#> 
#> $links
#> $links$code_list
#> $links$code_list$href
#> [1] "https://api.beta.ons.gov.uk/v1/code-lists/quarter"
#> 
#> 
#> $links$datasets
#> $links$datasets$href
#> [1] "https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q2/datasets"
#> 
#> 
#> $links$self
#> $links$self$href
#> [1] "https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q2"
#> 
#> $links$self$id
#> [1] "q2"
#> 
#> 
#> 

ons_code_dataset(id = "quarter", edition = "one-off", code = "q2")

#> list()

# }