Explore codes and lists

Used to get details about codes and code lists stored by ONS. Codes are used to provide a common definition when presenting statistics with related categories. Codes are gathered in code lists, which may change over time to include new or different codes. The meaning of a code should not change over time, but new codes may be created where new meaning is required.

ons_codelists()

ons_codelist(id = NULL, edition = NULL, code = NULL)

ons_codelist2(id = NULL, edition = NULL, code = NULL)

ons_codelist3(id = NULL, edition = NULL, code = NULL)

ons_codelist4(id = NULL, edition = NULL, code = NULL)

Examples

ons_codelists()

#> Error in ons_codelists(): could not find function "ons_codelists"

ons_codelist(id = "quarter")

#> Error in ons_codelist(id = "quarter"): could not find function "ons_codelist"

#editions
ons_codelist2(id = "quarter", edition = EMPTY)

#> Error in ons_codelist2(id = "quarter", edition = EMPTY): could not find function "ons_codelist2"

ons_codelist2(id = "quarter", edition = "one-off")

#> Error in ons_codelist2(id = "quarter", edition = "one-off"): could not find function "ons_codelist2"

#codes
ons_codelist3(id = "quarter", edition = "one-off", code = EMPTY)

#> Error in ons_codelist3(id = "quarter", edition = "one-off", code = EMPTY): could not find function "ons_codelist3"

ons_codelist3(id = "quarter", edition = "one-off", code = "q2")

#> Error in ons_codelist3(id = "quarter", edition = "one-off", code = "q2"): could not find function "ons_codelist3"

ons_codelist4(id = "quarter", edition = "one-off", code = "q2")

#> Error in ons_codelist4(id = "quarter", edition = "one-off", code = "q2"): could not find function "ons_codelist4"