Explore codes and lists

Used to get details about codes and code lists stored by ONS. Codes are used to provide a common definition when presenting statistics with related categories. Codes are gathered in code lists, which may change over time to include new or different codes. The meaning of a code should not change over time, but new codes may be created where new meaning is required.

ons_codes(id = NULL, edition = NULL)

Arguments

id
[character] Id that represents a dataset.

edition
[character] A subset of the dataset representing a specific time period. For some datasets this edition can contain all time periods (all historical data). The latest version of this is displayed by default.

code
[character] The ID of the code within a code list.

Examples

ons_codelists()

#>  [1] "administrative-geography"        "adzuna-jobs-category"           
#>  [3] "age-groups"                      "averages-and-percentiles"       
#>  [5] "build-status"                    "calendar-month"                 
#>  [7] "calendar-years"                  "cause-of-death"                 
#>  [9] "city-regions"                    "construction-classifications"   
#> [11] "construction-series-type"        "countries"                      
#> [13] "countries-and-territories"       "cpih1dim1aggid"                 
#> [15] "dd-mm"                           "decade"                         
#> [17] "economic-activity"               "england-regions"                
#> [19] "enterprise-regions"              "financial-and-calendar-years"   
#> [21] "healthcare-financing-scheme"     "healthcare-function"            
#> [23] "healthcare-provider"             "hours-and-earnings"             
#> [25] "house-sales-and-prices"          "household-reference-person"     
#> [27] "income-type"                     "index-and-year-change"          
#> [29] "life-expectancy-cohort"          "life-expectancy-type"           
#> [31] "local-enterprise-partnership"    "local-health-board"             
#> [33] "measure-of-wellbeing"            "middle-layer-super-output-area" 
#> [35] "mmm"                             "mmm-mmm-yyyy"                   
#> [37] "mmm-yy"                          "nuts"                           
#> [39] "parliamentary-constituencies"    "pedestrians-and-vehicles"       
#> [41] "place-of-death"                  "property-type"                  
#> [43] "quarter"                         "quarterly-index-and-growth-rate"
#> [45] "quintile"                        "recorded-deaths"                
#> [47] "registration-or-occurrence"      "seasonal-adjustment"            
#> [49] "sector"                          "sex"                            
#> [51] "ship-and-visit-type"             "shipping-port"                  
#> [53] "sic"                             "sic-unofficial"                 
#> [55] "single-year-of-age"              "sitc"                           
#> [57] "soc"                             "speak-welsh"                    
#> [59] "tax-benefit-type"                "trade-direction"                
#> [61] "traffic-camera-area"             "travel-to-work-area"            
#> [63] "two-year-intervals"              "type-of-prices"                 
#> [65] "uk-business-unit"                "uk-only"                        
#> [67] "unit-of-measure"                 "value-deflation"                
#> [69] "week-number"                     "wellbeing-estimate"             
#> [71] "welsh-household-composition"     "working-pattern"                
#> [73] "workplace-or-residence"          "yyyy"                           
#> [75] "yyyy-qq"                         "yyyy-yy"                        

ons_codelist(id = "quarter")

#> Error in ons_codelist(id = "quarter"): could not find function "ons_codelist"

#editions
ons_codelist_editions(id = "quarter")

#> Error in ons_codelist_editions(id = "quarter"): could not find function "ons_codelist_editions"

ons_codelist_edition(id = "quarter", edition = "one-off")

#> Error in ons_codelist_edition(id = "quarter", edition = "one-off"): could not find function "ons_codelist_edition"

#codes
ons_codes(id = "quarter", edition = "one-off")

#> $count
#> [1] 4
#> 
#> $items
#>   code label                                        links.href
#> 1   q1    Q1 https://api.beta.ons.gov.uk/v1/code-lists/quarter
#> 2   q2    Q2 https://api.beta.ons.gov.uk/v1/code-lists/quarter
#> 3   q3    Q3 https://api.beta.ons.gov.uk/v1/code-lists/quarter
#> 4   q4    Q4 https://api.beta.ons.gov.uk/v1/code-lists/quarter
#>                                                                             links.href
#> 1 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q1/datasets
#> 2 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q2/datasets
#> 3 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q3/datasets
#> 4 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q4/datasets
#>                                                               links.self.href
#> 1 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q1
#> 2 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q2
#> 3 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q3
#> 4 https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q4
#>   links.self.id
#> 1            q1
#> 2            q2
#> 3            q3
#> 4            q4
#> 
#> $limit
#> [1] 20
#> 
#> $offset
#> [1] 0
#> 
#> $total_count
#> [1] 4
#> 

ons_code(id = "quarter", edition = "one-off", code = "q2")

#> $code
#> [1] "q2"
#> 
#> $label
#> [1] ""
#> 
#> $links
#> $links$code_list
#> $links$code_list$href
#> [1] "https://api.beta.ons.gov.uk/v1/code-lists/quarter"
#> 
#> 
#> $links$datasets
#> $links$datasets$href
#> [1] "https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q2/datasets"
#> 
#> 
#> $links$self
#> $links$self$href
#> [1] "https://api.beta.ons.gov.uk/v1/code-lists/quarter/editions/one-off/codes/q2"
#> 
#> $links$self$id
#> [1] "q2"
#> 
#> 
#> 

ons_code_dataset(id = "quarter", edition = "one-off", code = "q2")

#> $count
#> [1] 0
#> 
#> $items
#> list()
#> 
#> $limit
#> [1] 20
#> 
#> $offset
#> [1] 0
#> 
#> $total_count
#> [1] 0
#>