ONS Datasets

A grouping of data (editions) with shared dimensions, for example Sex, Age and Geography, and all published history of this group of data. The options in these dimensions can change over time leading to separate editions. For example: Population Estimates for UK, England and Wales, Scotland and Northern Ireland.

ons_datasets()

ons_ids()

Value

A tibble with the datasets.

Examples

# \donttest{
# Find all the information about the data
ons_datasets()

#> # A tibble: 41 x 17
#>    contacts  description  id    keywords links$editions$… $latest_version… $$id 
#>    <list>    <chr>        <chr> <list>   <chr>            <chr>            <chr>
#>  1 <df[,3] … "Indicators… agei… <chr [2… https://api.bet… https://api.bet… 1    
#>  2 <df[,3] … "Projected … agei… <chr [3… https://api.bet… https://api.bet… 1    
#>  3 <df[,3] … "Indicators… olde… <chr [2… https://api.bet… https://api.bet… 1    
#>  4 <df[,3] … "Figures pr… olde… <chr [2… https://api.bet… https://api.bet… 1    
#>  5 <df[,3] … "Both indic… olde… <chr [2… https://api.bet… https://api.bet… 1    
#>  6 <df[,3] … "Projected … proj… <NULL>   https://api.bet… https://api.bet… 2    
#>  7 <df[,3] … "Projected … proj… <NULL>   https://api.bet… https://api.bet… 1    
#>  8 <df[,3] … "Provisiona… week… <chr [2… https://api.bet… https://api.bet… 25   
#>  9 <df[,3] … "Provisiona… week… <chr [2… https://api.bet… https://api.bet… 25   
#> 10 <df[,3] … "Experiment… onli… <NULL>   https://api.bet… https://api.bet… 4    
#> # … with 31 more rows, and 12 more variables: next_release <chr>,
#> #   qmi <tibble[,1]>, related_datasets <list>, release_frequency <chr>,
#> #   state <chr>, title <chr>, methodologies <list>, national_statistic <lgl>,
#> #   publications <list>, license <chr>, unit_of_measure <chr>, type <chr>

# Just the ids
ons_ids()

#>  [1] "ageing-population-estimates"                  
#>  [2] "ageing-population-projections"                
#>  [3] "older-people-economic-activity"               
#>  [4] "older-people-net-internal-migration"          
#>  [5] "older-people-sex-ratios"                      
#>  [6] "projections-older-people-in-single-households"
#>  [7] "projections-older-people-sex-ratios"          
#>  [8] "weekly-deaths-age-sex"                        
#>  [9] "weekly-deaths-region"                         
#> [10] "online-job-advert-estimates"                  
#> [11] "faster-indicators-shipping-data"              
#> [12] "cpih01"                                       
#> [13] "mid-year-pop-est"                             
#> [14] "weekly-deaths-health-board"                   
#> [15] "weekly-deaths-local-authority"                
#> [16] "regional-gdp-by-year"                         
#> [17] "wellbeing-local-authority"                    
#> [18] "index-private-housing-rental-prices"          
#> [19] "suicides-in-the-uk"                           
#> [20] "childrens-wellbeing"                          
#> [21] "gdp-to-four-decimal-places"                   
#> [22] "generational-income"                          
#> [23] "health-accounts"                              
#> [24] "wellbeing-quarterly"                          
#> [25] "house-prices-local-authority"                 
#> [26] "regional-gdp-by-quarter"                      
#> [27] "labour-market"                                
#> [28] "tax-benefits-statistics"                      
#> [29] "ashe-tables-26"                               
#> [30] "ashe-tables-25"                               
#> [31] "ashe-tables-27-and-28"                        
#> [32] "trade"                                        
#> [33] "ashe-tables-7-and-8"                          
#> [34] "ashe-tables-3"                                
#> [35] "ashe-tables-11-and-12"                        
#> [36] "ashe-tables-20"                               
#> [37] "ashe-tables-9-and-10"                         
#> [38] "ashe-table-5"                                 
#> [39] "life-expectancy-by-local-authority"           
#> [40] "gdp-by-local-authority"                       
#> [41] "gva-by-industry-by-local-authority"           

# }