Download data from ONS

This functions is used to find information about data published by the ONS. Datasets are published in unique versions, which are categorized by edition. Available datasets are given an id. All available id can be viewed with ons_ids().

ons_get(
  id = NULL,
  edition = NULL,
  version = NULL,
  ons_read = getOption("onsr.read"),
  ...
)

Arguments

id
[character] Id that represents a dataset.

edition
[character] A subset of the dataset representing a specific time period. For some datasets this edition can contain all time periods (all historical data). The latest version of this is displayed by default.

version
[character] A specific instance of the edition at a point in time. New versions can be published as a result of corrections, revisions or new data becoming available.

ons_read
[character]. Reading backend, one of readr, data.table or vroom.

...
Further arguments passed on the reading functions.

Value

A tibble with the dataset in tidy format.

Details

Examples

ons_get(id = "cpih01", limit = 500)

#> Error in readr::read_csv(x, ...): unused argument (limit = 500)

# Same dataset but older version
ons_get(id = "cpih01", version = "5")

#> # A tibble: 55,977 x 7
#>     v4_0 `mmm-yy` Time   `uk-only` Geography  cpih1dim1aggid Aggregate          
#>    <dbl> <chr>    <chr>  <chr>     <chr>      <chr>          <chr>              
#>  1  90.9 Jul-11   Jul-11 K02000001 United Ki… cpih1dim1G110… 11.1 Catering serv…
#>  2  68.7 Feb-98   Feb-98 K02000001 United Ki… cpih1dim1S101… 01.1.2 Meat        
#>  3  94.9 Mar-12   Mar-12 K02000001 United Ki… cpih1dim1S802… 08.2.0 Telephone a…
#>  4  84.9 Aug-93   Aug-93 K02000001 United Ki… cpih1dim1S120… 12.1.2/3 Appliance…
#>  5 102.  Dec-12   Dec-12 K02000001 United Ki… cpih1dim1S502… 05.2.0 Household T…
#>  6  91.4 Apr-91   Apr-91 K02000001 United Ki… cpih1dim1S903… 09.3.3 Gardens, pl…
#>  7   0   Jun-96   Jun-96 K02000001 United Ki… cpih1dim1S120… 12.4.0 Social Prot…
#>  8 108.  Mar-03   Mar-03 K02000001 United Ki… cpih1dim1S502… 05.2.0 Household T…
#>  9  30.8 May-96   May-96 K02000001 United Ki… cpih1dim1G202… 02.2 Tobacco       
#> 10  85.3 Jun-99   Jun-99 K02000001 United Ki… cpih1dim1G505… 05.5 Tools and equ…
#> # … with 55,967 more rows